Pull testing code out into a dedicated object
